Extract AuthenticationContext bean from controllers
Replace duplicated extractOrganizationId() and extractUserId()
private methods across 5 controllers with a shared injectable
AuthenticationContext component. Update controller tests with
mock bean stubs.
