+++ src/test/kotlin/com/froilan/synectix/controller/AccountControllerTest.kt
@@ -12,6 +12,7 @@ import com.froilan.synectix.repository.PasswordResetTokenRepository
 import com.froilan.synectix.repository.RefreshTokenRepository
 import com.froilan.synectix.repository.SessionTokenRepository
 import com.froilan.synectix.repository.UserRepository
+import com.froilan.synectix.security.AuthenticationContext
 import com.froilan.synectix.security.RolePermissionCache
 import com.froilan.synectix.security.SessionContext
 import com.froilan.synectix.security.SynectixPermissionEvaluator
@@ -85,6 +86,9 @@ class AccountControllerTest {
     @MockitoBean
     private lateinit var journalEntryService: JournalEntryService
 
+    @MockitoBean
+    private lateinit var authenticationContext: AuthenticationContext
+
     private val testUser =
         User(
             uuid = "user-123",
@@ -100,6 +104,8 @@ class AccountControllerTest {
     @BeforeEach
     fun setup() {
         setupAuthWithPermissions("account:create", "account:read", "account:update", "account:delete")
+        `when`(authenticationContext.organizationId()).thenReturn("org-123")
+        `when`(authenticationContext.userId()).thenReturn("user-123")
     }
 
     private fun setupAuthWithPermissions(vararg permissions: String) {
